# swagger 2/Servers/javascript/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "A sample API definition for a honeybee server"
  version: "0.0.0"
  title: "Honeybee Server API"
host: "api.ladybug-tools.io"
basePath: "/api"
schemes:
- "http"
paths:
  /recipe/:
    get:
      tags:
      - "Get and Delete Recipes"
      summary: "Get a list of daylight recipe objects"
      description: "Retrieves a list of daylight recipe objects for a given filter."
      operationId: "recipeGET"
      parameters:
      - name: "size"
        in: "query"
        description: "Number of files returned"
        required: false
        type: "integer"
      - name: "country"
        in: "query"
        description: "Country where definitions should be from"
        required: false
        type: "string"
      responses:
        200:
          description: "A list of analysis_grids matching the query parameters"
          schema:
            $ref: "#/definitions/inline_response_200"
        404:
          description: "Couldnt find any definitions matching your search query...\
            \ :("
      x-swagger-router-controller: "GetAndDeleteRecipes"
  /recipe/{uuid}:
    get:
      tags:
      - "Get and Delete Recipes"
      summary: "Get a specific analysis_grid file"
      description: "Returns a single analysis_grid file"
      operationId: "recipeUuidGET"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The unique identifier of the analysis_grid."
        required: true
        type: "string"
      responses:
        200:
          description: "Succesfully returns the recipe"
          schema:
            $ref: "#/definitions/GridBasedRecipeSchema"
        400:
          description: "The uuid provided is invalid"
        404:
          description: "The analysis_grid does not exist"
      x-swagger-router-controller: "GetAndDeleteRecipes"
    delete:
      tags:
      - "Get and Delete Recipes"
      summary: "Delete an existing recipe object"
      description: "Delete an recipe object in the database specifying the uuid of\
        \ the file"
      operationId: "recipeUuidDELETE"
      parameters:
      - name: "uuid"
        in: "path"
        description: "The unique identifier of the recipe."
        required: true
        type: "string"
      responses:
        200:
          description: "File succesfully deleted"
        400:
          description: "Bad request, something went wrong... :("
          schema:
            $ref: "#/definitions/inline_response_400"
        404:
          description: "File not found"
      x-swagger-router-controller: "GetAndDeleteRecipes"
...
definitions:
...
  GridBasedRecipeSchema:
    properties:
      uuid:
        type: "string"
        description: "The unique identifier of the recipe"
    example:
      uuid: "709116b8-d71e-4d50-bb54-216e83ad9f3f"
      recipe:
        id: "annual"
        type: "gridbased"
        sky_mtx:
          wea: "wea Json file"
          sky_density: 1
          north: 0
          mode: 2
          suffix: null
        analysis_grids:
        - "list of analysis grids"
        surfaces:
        - "list of surfaces"
        analysis_type: 2
        rad_parameters: "-ab 5 -aa 0.05 -ar 128"
...
  inline_response_200:
    properties:
      query:
        type: "object"
        properties: {}
      results:
        type: "array"
        items:
          $ref: "#/definitions/GridBasedRecipeSchema"
    example:
      query: "{}"
      results:
      - uuid: "709116b8-d71e-4d50-bb54-216e83ad9f3f"
        recipe:
          id: "annual"
          type: "gridbased"
          sky_mtx:
            wea: "wea Json file"
            sky_density: 1
            north: 0
            mode: 2
            suffix: null
          analysis_grids:
          - "list of analysis grids"
          surfaces:
          - "list of surfaces"
          analysis_type: 2
          rad_parameters: "-ab 5 -aa 0.05 -ar 128"
      - uuid: "709116b8-d71e-4d50-bb54-216e83ad9f3f"
        recipe:
          id: "annual"
          type: "gridbased"
          sky_mtx:
            wea: "wea Json file"
            sky_density: 1
            north: 0
            mode: 2
            suffix: null
          analysis_grids:
          - "list of analysis grids"
          surfaces:
          - "list of surfaces"
          analysis_type: 2
          rad_parameters: "-ab 5 -aa 0.05 -ar 128"
  inline_response_400:
    properties:
      error:
        type: "string"
...